Remove lib directory.
First step of re-organising files to use a more typical structure.
At some stage I wanted to try to write client-side 'components' as small packages of HTML, CSS and JavaScript and I thought a non-standard structure might help to demonstrate that. We decided not to do that in the end anyway, so moving to a more typical file structure makes sense.